Split Lean vocabulary test registry
