+++ crates/defra-agent/src/lean_vocab_test.rs
@@ -157,712 +157,25 @@ pub(crate) struct LeanLifecycleTransitionCase {
     pub(crate) boundary: Option<String>,
 }
 
-#[derive(Debug, Deserialize, Clone, PartialEq)]
-pub(crate) struct LeanEventDeliveryWorld {
-    pub(crate) persistent_set: Vec<String>,
-    pub(crate) subscription_queue: Vec<String>,
-    pub(crate) processed_set: Vec<String>,
-    pub(crate) handled: Vec<String>,
-}
+#[path = "lean_vocab_test/background_transcript.rs"]
+mod background_transcript;
+#[path = "lean_vocab_test/client_session.rs"]
+mod client_session;
+#[path = "lean_vocab_test/command_identity_queue.rs"]
+mod command_identity_queue;
+#[path = "lean_vocab_test/event_delivery.rs"]
+mod event_delivery;
+#[path = "lean_vocab_test/slot_persistence_health.rs"]
+mod slot_persistence_health;
+#[path = "lean_vocab_test/triggers_runtime_apply.rs"]
+mod triggers_runtime_apply;
 
-#[derive(Debug, Deserialize, Clone, PartialEq)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub(crate) enum LeanEventDeliveryAction {
-    Persist { doc: String },
-    Depersist { doc: String },
-    Enqueue { doc: String },
-    Drop { doc: String },
-    DeliverFromQueue { doc: String },
-    RescanTick,
-    Handle { doc: String },
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub(crate) struct LeanEventDeliveryTransitionCase {
-    pub(crate) name: String,
-    pub(crate) pre: LeanEventDeliveryWorld,
-    pub(crate) action: LeanEventDeliveryAction,
-    pub(crate) post: LeanEventDeliveryWorld,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub(crate) struct LeanEventDeliverySourceInstance {
-    pub(crate) name: String,
-    pub(crate) dedupe_policy: String,
-    pub(crate) rescan_bounded_by: u64,
-    pub(crate) deviation: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub(crate) struct LeanEventDeliveryConvergenceTrace {
-    pub(crate) name: String,
-    pub(crate) instance_name: String,
-    pub(crate) initial_world: LeanEventDeliveryWorld,
-    pub(crate) actions: Vec<LeanEventDeliveryAction>,
-    pub(crate) final_world: LeanEventDeliveryWorld,
-    pub(crate) status: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanTriggerKeyContract {
-    pub(crate) trigger_id: String,
-    pub(crate) trigger_kind: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanTriggerDispatchCase {
-    pub(crate) name: String,
-    pub(crate) trigger_id: Option<String>,
-    pub(crate) trigger_kind: String,
-    pub(crate) concurrency: String,
-    pub(crate) active_schedule_ids: Vec<String>,
-    pub(crate) active_event_trigger_ids: Vec<String>,
-    pub(crate) prior_nonterminal_keys: Vec<LeanTriggerKeyContract>,
-    pub(crate) expected_result: String,
-    pub(crate) expected_skip_reason: Option<String>,
-    pub(crate) expected_materialize_trigger_id: Option<String>,
-    pub(crate) expected_materialize_trigger_kind: Option<String>,
-    pub(crate) expected_request_caused_by_id: Option<String>,
-    pub(crate) expected_request_caused_by_kind: Option<String>,
-    pub(crate) expected_execution_origin: Option<String>,
-    pub(crate) expected_supersede_call_keys: Vec<LeanTriggerKeyContract>,
-    pub(crate) superseded_prior_ids: Vec<String>,
-    pub(crate) target_nonterminal_count_after: Option<usize>,
-    pub(crate) request_count_before: usize,
-    pub(crate) request_count_after: usize,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanRuntimeReconcileCase {
-    pub(crate) name: String,
-    pub(crate) action: String,
-    pub(crate) legal: bool,
-    pub(crate) pre_phase: String,
-    pub(crate) post_phase: String,
-    pub(crate) pre_active_generation: usize,
-    pub(crate) post_active_generation: usize,
-    pub(crate) pre_router_generation: usize,
-    pub(crate) post_router_generation: usize,
-    pub(crate) pre_ready_generation_count: usize,
-    pub(crate) post_ready_generation_count: usize,
-    pub(crate) pre_live_generation_count: usize,
-    pub(crate) post_live_generation_count: usize,
-    pub(crate) pre_in_flight_count: usize,
-    pub(crate) post_in_flight_count: usize,
-    pub(crate) tracked_request_id: usize,
-    pub(crate) tracked_session_id: usize,
-    pub(crate) tracked_request_generation: usize,
-    pub(crate) tracked_request_session: usize,
-    pub(crate) tracked_request_behavior: usize,
-    pub(crate) tracked_session_behavior: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanApplyDocRef {
-    pub(crate) collection: String,
-    pub(crate) id: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanApplyDesiredDoc {
-    pub(crate) collection: String,
-    pub(crate) id: String,
-    pub(crate) content: String,
-    pub(crate) refs: Vec<LeanApplyDocRef>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanApplyLiveDoc {
-    pub(crate) collection: String,
-    pub(crate) id: String,
-    pub(crate) content: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanApplyStep {
-    pub(crate) action: String,
-    pub(crate) target: LeanApplyDocRef,
-    pub(crate) content: String,
-    pub(crate) refs: Vec<LeanApplyDocRef>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanApplyCollectionWrite {
-    pub(crate) collection: String,
-    pub(crate) graphql_type: String,
-    pub(crate) unique_field: String,
-    pub(crate) apply_order: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanApplySelectedDoc {
-    pub(crate) action: String,
-    pub(crate) target: LeanApplyDocRef,
-    pub(crate) graphql_type: String,
-    pub(crate) unique_field: String,
-    pub(crate) unique_value: String,
-    pub(crate) content: String,
-    pub(crate) refs: Vec<LeanApplyDocRef>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanApplyReconcileCase {
-    pub(crate) name: String,
-    pub(crate) manifest: Vec<LeanApplyDesiredDoc>,
-    pub(crate) pre_desired: Vec<LeanApplyDesiredDoc>,
-    pub(crate) pre_live: Vec<LeanApplyLiveDoc>,
-    pub(crate) expected_create: Vec<LeanApplyDocRef>,
-    pub(crate) expected_update: Vec<LeanApplyDocRef>,
-    pub(crate) expected_unchanged: Vec<LeanApplyDocRef>,
-    pub(crate) expected_live_only: Vec<LeanApplyDocRef>,
-    pub(crate) expected_steps: Vec<LeanApplyStep>,
-    pub(crate) expected_write_order: Vec<LeanApplyCollectionWrite>,
-    pub(crate) expected_selected_create_docs: Vec<LeanApplySelectedDoc>,
-    pub(crate) expected_selected_update_docs: Vec<LeanApplySelectedDoc>,
-    pub(crate) expected_selected_writes: Vec<LeanApplySelectedDoc>,
-    pub(crate) prefix_len: usize,
-    pub(crate) expected_prefix_desired: Vec<LeanApplyDesiredDoc>,
-    pub(crate) expected_after_desired: Vec<LeanApplyDesiredDoc>,
-    pub(crate) expected_retry_desired: Vec<LeanApplyDesiredDoc>,
-    pub(crate) expected_retry_step_count: usize,
-    pub(crate) expected_rediff_step_count: usize,
-    pub(crate) live_preserved: bool,
-    pub(crate) manifest_realized_after: bool,
-    pub(crate) retry_converges: bool,
-    pub(crate) idempotent_after: bool,
-    pub(crate) write_order_prefix_safe: bool,
-    pub(crate) production_prefixes_referrers_closed: bool,
-    pub(crate) prefix_referrers_closed: bool,
-    pub(crate) desired_references_closed_after_prefix: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanClientShellCase {
-    pub(crate) name: String,
-    pub(crate) property: String,
-    pub(crate) input: String,
-    pub(crate) pre_selection_session: Option<usize>,
-    pub(crate) post_selection_session: Option<usize>,
-    pub(crate) pre_workflow_kind: String,
-    pub(crate) pre_workflow_request: Option<usize>,
-    pub(crate) post_workflow_kind: String,
-    pub(crate) post_workflow_request: Option<usize>,
-    pub(crate) selection_preserved: bool,
-    pub(crate) workflow_advanced: bool,
-    pub(crate) transport_noop: bool,
-    pub(crate) can_submit_before: bool,
-    pub(crate) can_submit_after: bool,
-    pub(crate) send_decision: String,
-    pub(crate) send_blocked_reason: Option<String>,
-    pub(crate) frontend_expected_workflow_kind: String,
-    pub(crate) frontend_expected_send_status: String,
-    pub(crate) frontend_expected_send_blocked_reason: Option<String>,
-    pub(crate) frontend_expected_active_request_id: Option<usize>,
-    pub(crate) frontend_conversation_present: bool,
-    pub(crate) desktop_selected_session_id: Option<usize>,
-    pub(crate) desktop_snapshot_present: bool,
-    pub(crate) desktop_preferred_request_id: Option<usize>,
-    pub(crate) desktop_observed_request_id: Option<usize>,
-    pub(crate) desktop_observed_turn_state: Option<String>,
-    pub(crate) desktop_expected_latest_request_id: Option<usize>,
-    pub(crate) desktop_expected_turn_state: Option<String>,
-    pub(crate) desktop_expect_pending_turn: Option<bool>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanSessionRecoveryCase {
-    pub(crate) name: String,
-    pub(crate) action: String,
-    pub(crate) legal: bool,
-    pub(crate) pre_latest_state: String,
-    pub(crate) pre_failed_state: String,
-    pub(crate) post_latest_state: String,
-    pub(crate) post_failed_state: String,
-    pub(crate) post_new_state: String,
-    pub(crate) pre_latest_admission: String,
-    pub(crate) post_latest_admission: String,
-    pub(crate) pre_failed_admission: String,
-    pub(crate) post_failed_admission: String,
-    pub(crate) post_new_admission: String,
-    pub(crate) pre_origin: String,
-    pub(crate) post_new_origin: String,
-    pub(crate) pre_backend: String,
-    pub(crate) post_new_backend: String,
-    pub(crate) failed_id: usize,
-    pub(crate) new_id: usize,
-    pub(crate) pre_latest_id: usize,
-    pub(crate) post_latest_id: usize,
-    pub(crate) pre_session_id: usize,
-    pub(crate) post_session_id: usize,
-    pub(crate) pre_behavior_id: usize,
-    pub(crate) post_behavior_id: usize,
-    pub(crate) pre_request_count: usize,
-    pub(crate) post_request_count: usize,
-    pub(crate) pre_retry_count: usize,
-    pub(crate) post_retry_count: usize,
-    pub(crate) max_retries: usize,
-    pub(crate) pre_deadline_exceeded: bool,
-    pub(crate) post_deadline_exceeded: bool,
-    pub(crate) pre_failed_is_latest: bool,
-    pub(crate) post_failed_is_latest: bool,
-    pub(crate) post_new_is_latest: bool,
-    pub(crate) pre_request_ids: Vec<usize>,
-    pub(crate) pre_failed_exists: bool,
-    pub(crate) pre_latest_exists: bool,
-    pub(crate) pre_new_request_exists: bool,
-    pub(crate) old_request_retained: bool,
-    pub(crate) new_request_inserted: bool,
-    pub(crate) origin_preserved: bool,
-    pub(crate) backend_preserved: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanInferenceSlotAccountingCase {
-    pub(crate) name: String,
-    pub(crate) property: String,
-    pub(crate) backend_id: String,
-    pub(crate) pre_state: String,
-    pub(crate) post_state: String,
-    pub(crate) contribution: usize,
-    pub(crate) expected_contribution: usize,
-    pub(crate) pre_contribution: usize,
-    pub(crate) post_contribution: usize,
-    pub(crate) released_slot: bool,
-    pub(crate) permit_drop_terminalization: bool,
-    pub(crate) row_states: Vec<String>,
-    pub(crate) row_backend_ids: Vec<String>,
-    pub(crate) reconstructed_running_count: usize,
-    pub(crate) max_concurrent: usize,
-    pub(crate) bounded_by_max_concurrent: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanFleetSlotAccountingCase {
-    pub(crate) name: String,
-    pub(crate) property: String,
-    pub(crate) backend_id: String,
-    pub(crate) request_state: String,
-    pub(crate) admission_state: String,
-    pub(crate) contribution: usize,
-    pub(crate) expected_contribution: usize,
-    pub(crate) active_count: usize,
-    pub(crate) scheduler_running: usize,
-    pub(crate) slot_count: usize,
-    pub(crate) row_states: Vec<String>,
-    pub(crate) row_backend_ids: Vec<String>,
-    pub(crate) reconstructed_running_count: usize,
-    pub(crate) max_concurrent: usize,
-    pub(crate) bounded_by_max_concurrent: bool,
-    pub(crate) aggregate_reconstructed_not_persisted: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanPersistenceFailurePolicyCase {
-    pub(crate) name: String,
-    pub(crate) policy: String,
-    pub(crate) action: String,
-    pub(crate) pre_persistence: String,
-    pub(crate) post_persistence: String,
-    pub(crate) post_storage_observation: String,
-    pub(crate) hook_decision: String,
-    pub(crate) records_failure: bool,
-    pub(crate) records_success: bool,
-    pub(crate) external_durability_claimed: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanStorageObservationRuntimeCase {
-    pub(crate) name: String,
-    pub(crate) policy: String,
-    pub(crate) action: String,
-    pub(crate) pre_observation: String,
-    pub(crate) mutation_result: String,
-    pub(crate) post_observation: String,
-    pub(crate) post_persistence: String,
-    pub(crate) hook_result: String,
-    pub(crate) records_failure: bool,
-    pub(crate) records_success: bool,
-    pub(crate) terminal_write_observed: bool,
-    pub(crate) external_visibility_claimed: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanBackendHealthAdmissionCase {
-    pub(crate) name: String,
-    pub(crate) enabled: bool,
-    pub(crate) probe_status: String,
-    pub(crate) expected_available: bool,
-    pub(crate) admission_decision: String,
-    pub(crate) observed_document_only: bool,
-    pub(crate) external_endpoint_freshness_claimed: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanNativeFilesystemBoundaryCase {
-    pub(crate) name: String,
-    pub(crate) tool_name: String,
-    pub(crate) work_class: String,
-    pub(crate) boundary: String,
-    pub(crate) inner_poll_blocks: bool,
-    pub(crate) request_deadline_ms: usize,
-    pub(crate) blocker_ms: usize,
-    pub(crate) expected_terminal: String,
-    pub(crate) expected_failure_class: Option<String>,
-    pub(crate) queue_advances_before_blocker_returns: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanToolPreflightCase {
-    pub(crate) name: String,
-    pub(crate) health: String,
-    pub(crate) schema_status: String,
-    pub(crate) decision: String,
-    pub(crate) failure_class: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanToolRetryCase {
-    pub(crate) name: String,
-    pub(crate) operation: String,
-    pub(crate) idempotency: String,
-    pub(crate) failure_class: String,
-    pub(crate) disposition: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanMcpHealthCase {
-    pub(crate) name: String,
-    pub(crate) start_state: String,
-    pub(crate) start_count: usize,
-    pub(crate) event: String,
-    pub(crate) threshold_k: usize,
-    pub(crate) next_state: Option<String>,
-    pub(crate) next_count: Option<usize>,
-    pub(crate) rust_projection: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanCommandPolicyCase {
-    pub(crate) name: String,
-    pub(crate) category: String,
-    pub(crate) mode: String,
-    pub(crate) allowed_argv_prefixes: Vec<Vec<String>>,
-    pub(crate) forbidden_argv_prefixes: Vec<Vec<String>>,
-    pub(crate) network_mode: String,
-    pub(crate) read_only_allowlist: Vec<String>,
-    pub(crate) command: String,
-    pub(crate) lookup_command: String,
-    pub(crate) args: Vec<String>,
-    pub(crate) decision: String,
-    pub(crate) denial_reason: Option<String>,
-    pub(crate) matched_prefix: Option<Vec<String>>,
-    pub(crate) denied_argv: Option<Vec<String>>,
-    pub(crate) denied_command: Option<String>,
-    pub(crate) denied_argument: Option<String>,
-    pub(crate) denied_subcommand: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanCommandSandboxCase {
-    pub(crate) name: String,
-    pub(crate) category: String,
-    pub(crate) mode: String,
-    pub(crate) workspace_write_sandbox_enforced: bool,
-    pub(crate) decision: String,
-    pub(crate) sandbox: Option<String>,
-    pub(crate) denial_reason: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanCommandEnvCase {
-    pub(crate) name: String,
-    pub(crate) env_key: String,
-    pub(crate) input_present: bool,
-    pub(crate) input_name: String,
-    pub(crate) input_value: String,
-    pub(crate) output_name: String,
-    pub(crate) expected_value_kind: Option<String>,
-    pub(crate) expected_output_value: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanLiveOverlayCase {
-    pub(crate) name: String,
-    #[serde(rename = "responseStatus")]
-    pub(crate) response_status: String,
-    pub(crate) materialized: bool,
-    #[serde(rename = "precedingToolCalls")]
-    pub(crate) preceding_tool_calls: u64,
-    #[serde(rename = "turnTerminal")]
-    pub(crate) turn_terminal: bool,
-    #[serde(rename = "turnLabel")]
-    pub(crate) turn_label: String,
-    #[serde(rename = "hasContent")]
-    pub(crate) has_content: bool,
-    #[serde(rename = "hasReasoning")]
-    pub(crate) has_reasoning: bool,
-    #[serde(rename = "expectOverlay")]
-    pub(crate) expect_overlay: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityPrincipal {
-    pub(crate) did: String,
-    pub(crate) enabled: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityBehavior {
-    pub(crate) id: String,
-    pub(crate) principal: String,
-    pub(crate) enabled: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityDeployment {
-    pub(crate) id: String,
-    pub(crate) principal: String,
-    pub(crate) host_id: String,
-    pub(crate) enabled: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityStructuralCase {
-    pub(crate) name: String,
-    pub(crate) principals: Vec<LeanIdentityPrincipal>,
-    pub(crate) behaviors: Vec<LeanIdentityBehavior>,
-    pub(crate) deployments: Vec<LeanIdentityDeployment>,
-    pub(crate) well_formed: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityPermissionGrant {
-    pub(crate) principal: String,
-    pub(crate) permission: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityPermissionCase {
-    pub(crate) name: String,
-    pub(crate) principals: Vec<LeanIdentityPrincipal>,
-    pub(crate) behaviors: Vec<LeanIdentityBehavior>,
-    pub(crate) deployments: Vec<LeanIdentityDeployment>,
-    pub(crate) grants: Vec<LeanIdentityPermissionGrant>,
-    pub(crate) permission: String,
-    pub(crate) row_owner: String,
-    pub(crate) actor_behavior: String,
-    pub(crate) peer_behavior: String,
-    pub(crate) expected_actor_principal: String,
-    pub(crate) expected_peer_principal: String,
-    pub(crate) expected_actor_allowed: bool,
-    pub(crate) expected_peer_allowed: bool,
-    pub(crate) same_principal: bool,
-    pub(crate) expected_decisions_equal: bool,
-    pub(crate) host_deployment: String,
-    pub(crate) expected_actor_hostable: bool,
-    pub(crate) expected_peer_hostable: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanIdentityContract {
-    pub(crate) name: String,
-    pub(crate) statement: String,
-    pub(crate) enforced: bool,
-    pub(crate) tracked_by: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanQueueDeadlineConformanceCase {
-    pub(crate) name: String,
-    pub(crate) group: String,
-    pub(crate) action: String,
-    pub(crate) session_id: usize,
-    pub(crate) legal: bool,
-    pub(crate) pre_active_request_id: Option<usize>,
-    pub(crate) post_active_request_id: Option<usize>,
-    pub(crate) pre_pending_request_ids: Vec<usize>,
-    pub(crate) post_pending_request_ids: Vec<usize>,
-    pub(crate) claimed_request_id: Option<usize>,
-    pub(crate) blocked_by_active: bool,
-    pub(crate) superseded_request_ids: Vec<usize>,
-    pub(crate) queue_key: Option<String>,
-    pub(crate) post_coalesced_pending_count: usize,
-    pub(crate) automated_drained_request_ids: Vec<usize>,
-    pub(crate) preserved_user_pending_request_ids: Vec<usize>,
-    pub(crate) post_terminal_request_ids: Vec<usize>,
-    pub(crate) pre_request_deadline: Option<usize>,
-    pub(crate) synthesized_claim_deadline: Option<usize>,
-    pub(crate) post_deadline: Option<usize>,
-    pub(crate) explicit_deadline_preserved: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanRecoverySweepCase {
-    pub(crate) name: String,
-    pub(crate) sweep_id: String,
-    pub(crate) collection: String,
-    pub(crate) rust_function: String,
-    pub(crate) cadence: String,
-    pub(crate) implementation_status: String,
-    pub(crate) pre_state: String,
-    pub(crate) terminal_state: String,
-    pub(crate) measure_before: usize,
-    pub(crate) measure_after: usize,
-    pub(crate) deadline_audit_ref: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(tag = "witness", deny_unknown_fields)]
-pub(crate) enum LeanR4cBackgroundWorkCase {
-    #[serde(rename = "r4c.list_subagents.lineage_rejects")]
-    ListSubagentsLineageRejects {
-        caller_request_id: String,
-        sibling_request_id: String,
-        sibling_child_id: String,
-        caller_sees_sibling_child: bool,
-    },
-    #[serde(rename = "r4c.read_subagent_transcript.cursor_advances")]
-    ReadTranscriptCursorAdvances {
-        child_session_id: String,
-        first_since_sequence: usize,
-        first_through_sequence: usize,
-        first_next_sequence: usize,
-        second_since_sequence: usize,
-        second_through_sequence: usize,
-        no_gap: bool,
-        no_overlap: bool,
-    },
-    #[serde(rename = "r4c.read_subagent_transcript.hides_bridge_rows")]
-    ReadTranscriptHidesBridgeRows {
-        child_session_id: String,
-        bridge_call_id: String,
-        rendered_transcript: String,
-    },
-    #[serde(rename = "r4c.read_tool_output.dispatch_by_state")]
-    ReadToolOutputDispatchesByState {
-        tool_call_id: String,
-        running_source: String,
-        terminal_source: String,
-        running_payload: String,
-        stale_running_payload: String,
-        terminal_payload: String,
-    },
-    #[serde(rename = "r4c.steer_subagent.append_preserves_lineage")]
-    SteerAppendPreservesLineage {
-        caller_request_id: String,
-        child_session_id: String,
-        queued_request_id: String,
-        caused_by_parent_request_id: String,
-        queue_source: String,
-        queue_policy: String,
-    },
-    #[serde(rename = "r4c.steer_subagent.interrupt_composes")]
-    SteerInterruptComposes {
-        caller_request_id: String,
-        child_session_id: String,
-        interrupted_active_request_id: String,
-        drained_wake_up_request_ids: Vec<String>,
-        drained_wake_up_queue_key: String,
-        queued_request_id: String,
-        queue_interrupted_request_id: String,
-    },
-}
-
-impl LeanR4cBackgroundWorkCase {
-    pub(crate) fn witness(&self) -> &'static str {
-        match self {
-            Self::ListSubagentsLineageRejects { .. } => "r4c.list_subagents.lineage_rejects",
-            Self::ReadTranscriptCursorAdvances { .. } => {
-                "r4c.read_subagent_transcript.cursor_advances"
-            }
-            Self::ReadTranscriptHidesBridgeRows { .. } => {
-                "r4c.read_subagent_transcript.hides_bridge_rows"
-            }
-            Self::ReadToolOutputDispatchesByState { .. } => {
-                "r4c.read_tool_output.dispatch_by_state"
-            }
-            Self::SteerAppendPreservesLineage { .. } => {
-                "r4c.steer_subagent.append_preserves_lineage"
-            }
-            Self::SteerInterruptComposes { .. } => "r4c.steer_subagent.interrupt_composes",
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-pub(crate) struct LeanR6BackgroundingCase {
-    pub(crate) name: String,
-    pub(crate) group: String,
-    pub(crate) action: String,
-    pub(crate) legal: bool,
-    pub(crate) pre_live_count: usize,
-    pub(crate) max_backgrounded: usize,
-    pub(crate) await_mode: String,
-    pub(crate) cancel_policy: String,
-    pub(crate) child_request_id: Option<String>,
-    pub(crate) terminal_state: String,
-    pub(crate) result: Option<String>,
-    pub(crate) reason: Option<String>,
-    pub(crate) error_code: Option<String>,
-    pub(crate) queue_source: Option<String>,
-    pub(crate) queue_key: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanTranscriptCase {
-    pub(crate) name: String,
-    pub(crate) group: String,
-    pub(crate) action: String,
-    pub(crate) legal: bool,
-    pub(crate) pre_message_count: usize,
-    pub(crate) post_message_count: usize,
-    pub(crate) pre_tool_call_count: usize,
-    pub(crate) post_tool_call_count: usize,
-    pub(crate) pre_in_flight_count: usize,
-    pub(crate) post_in_flight_count: usize,
-    pub(crate) assistant_sequence: usize,
-    pub(crate) result_sequence: usize,
-    pub(crate) logical_result_id: usize,
-    pub(crate) payload_hash: usize,
-    pub(crate) expected_pair_closed: bool,
-    pub(crate) expected_ordered: bool,
-    pub(crate) expected_duplicate_reused_sequence: bool,
-    pub(crate) expected_strong_drain: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanResponseTransitionCase {
-    pub(crate) name: String,
-    pub(crate) group: String,
-    pub(crate) action: String,
-    pub(crate) legal: bool,
-    pub(crate) pre_status: String,
-    pub(crate) post_status: String,
-    pub(crate) pre_live_tail: String,
-    pub(crate) post_live_tail: String,
-    pub(crate) pre_token_count: usize,
-    pub(crate) post_token_count: usize,
-    pub(crate) error_reason: Option<String>,
-    pub(crate) pre_materialized_seq: Option<usize>,
-    pub(crate) post_materialized_seq: Option<usize>,
-    pub(crate) expected_request_state: Option<String>,
-    pub(crate) expected_request_persistence: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct LeanCompactionReducerCase {
-    pub(crate) name: String,
-    pub(crate) group: String,
-    pub(crate) reducer: String,
-    pub(crate) legal: bool,
-    pub(crate) pre_message_count: usize,
-    pub(crate) post_message_count: usize,
-    pub(crate) preserves_pairs: bool,
-    pub(crate) preserves_order: bool,
-    pub(crate) gate_open: bool,
-    pub(crate) safe_to_reduce: bool,
-    pub(crate) reducer_is_identity: bool,
-}
+pub(crate) use background_transcript::*;
+pub(crate) use client_session::*;
+pub(crate) use command_identity_queue::*;
+pub(crate) use event_delivery::*;
+pub(crate) use slot_persistence_health::*;
+pub(crate) use triggers_runtime_apply::*;
 
 #[derive(Debug, PartialEq, Eq)]
 enum LeanVocabularyParseError<'a> {
